add structure for tune pipeline
